test: 단위 테스트 분리
